+++ scripts/repeated_inference.py
@@ -21,5 +21,4 @@ def main(cfg: DictConfig) -> None:
 
 
 if __name__ == "__main__":
-
     main()