+++ tests/test_detector.py
@@ -9,7 +9,10 @@ import torch
 @pytest.mark.detector
 @pytest.mark.downloader
 def test_downloader_run(analyzer):
-    analyzer.detector.downloader.run()
+    if os.path.exists(analyzer.detector.downloader.path_local):
+        os.remove(analyzer.detector.downloader.path_local)
+    if not os.path.exists(analyzer.detector.downloader.path_local):
+        analyzer.detector.downloader.run()
     assert os.path.exists(analyzer.detector.downloader.path_local)